refactor: move OptionalDateTransform to utils and update imports
